requirements.txt
pymongo==4.6.1
python-dotenv==1.0.0
python-jose==3.3.0
bcrypt==4.1.2 